Fix failure to export checkboxes that default to true
This was because the game threw out all false settings when exporting -
in the past, every checkbox defaulted to false. This was done because
many checkboxes lack default settings of false, meaning their
defualt value was undefined, not false. The code now checks for this
case - false setting, undefined default - before moving on to the more
general check for equivalence.